Add an agent plugin to support elasticsearch7  (#6759)
