+++ docs/en/setup/service-agent/java-agent/Plugin-list.md
@@ -16,6 +16,7 @@
 - elastic-job-3.x
 - elasticsearch-5.x
 - elasticsearch-6.x
+- elasticsearch-7.x
 - feign-default-http-9.x
 - feign-pathvar-9.x
 - finagle
