+++ test/plugin/scenarios/elasticsearch-7.x-scenario/pom.xml
@@ -98,6 +98,18 @@
             <artifactId>elasticsearch</artifactId>
             <version>${test.framework.version}</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.elasticsearch.client</groupId>
+            <artifactId>transport</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.elasticsearch.plugin</groupId>
+            <artifactId>transport-netty4-client</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
